stacks/homelab_bot/requirements.txt
python-telegram-bot==20.7
httpx~=0.25.2
